feat(validator): moved validator outside of middleware
