+++ deno_dist/middleware.ts
@@ -11,5 +11,4 @@ export * from './middleware/jwt/index.ts'
 export * from './middleware/logger/index.ts'
 export * from './middleware/powered-by/index.ts'
 export * from './middleware/pretty-json/index.ts'
-export * from './middleware/validator/index.ts'
 export * from './adapter/deno/serve-static.ts'
