+++ deno_dist/validator/index.ts
@@ -1,0 +1,1 @@
+export { validator } from './validator.ts'
