+++ src/client/client.test.ts
@@ -5,8 +5,8 @@ import { rest } from 'msw'
 import { setupServer } from 'msw/node'
 import _fetch, { Request as NodeFetchRequest } from 'node-fetch'
 import { Hono } from '../hono'
-import { validator } from '../middleware/validator'
 import type { Equal, Expect } from '../utils/types'
+import { validator } from '../validator'
 import { hc } from './client'
 import type { InferResponseType } from './types'
