+++ src/middleware.ts
@@ -11,5 +11,4 @@ export * from './middleware/jwt'
 export * from './middleware/logger'
 export * from './middleware/powered-by'
 export * from './middleware/pretty-json'
-export * from './middleware/validator'
 export * from './adapter/deno/serve-static'
